ELSA1-387 Fjerner `<ScrollableContainer>`
Komponenten brukes ikke hos konsumenter og kan være problematisk med tanke på UU. Derfor slettes den.
I Elsa-komponenter som `<Drawer>` bruker vi vanlig scrollbar med CSS overflow istedet.
