+++ packages/components/src/components/Drawer/Drawer.module.css
@@ -1,4 +1,5 @@
-:root {
+.container {
+  /* Justering på spacing slik at fokusmarkering osv. ikke skjules ved overflow */
   --dds-drawer-content-container-padding: var(--dds-spacing-x0-25);
   --dds-drawer-container-padding: calc(
       var(--dds-spacing-x1) - var(--dds-drawer-content-container-padding)
@@ -6,9 +7,7 @@
     calc(var(--dds-spacing-x1) - var(--dds-drawer-content-container-padding))
     calc(var(--dds-spacing-x1) - var(--dds-drawer-content-container-padding))
     calc(var(--dds-spacing-x1-5) - var(--dds-drawer-content-container-padding));
-}
 
-.container {
   position: fixed;
   top: 0;
   height: 100%;
@@ -64,6 +63,7 @@
   display: flex;
   flex-direction: column;
   gap: var(--dds-spacing-x1);
+  /* Spacing slik at fokusmarkering osv. ikke skjules ved overflow */
   padding: var(--dds-drawer-content-container-padding);
 }
