+++ packages/components/src/components/Modal/Modal.module.css
@@ -25,17 +25,33 @@
 }
 
 .container {
+  /* Justering på spacing slik at fokusmarkering osv. ikke skjules ved overflow */
+  --dds-modal-content-padding: var(--dds-spacing-x0-25);
+  --dds-modal-container-padding: calc(
+      var(--dds-spacing-x1) - var(--dds-modal-content-padding)
+    )
+    calc(var(--dds-spacing-x1) - var(--dds-modal-content-padding))
+    calc(var(--dds-spacing-x1-5) - var(--dds-modal-content-padding))
+    calc(var(--dds-spacing-x1-5) - var(--dds-modal-content-padding));
   display: flex;
   flex-direction: column-reverse;
   min-width: 200px;
-  padding: var(--dds-spacing-x1) var(--dds-spacing-x1) var(--dds-spacing-x1-5)
-    var(--dds-spacing-x1-5);
+  padding: var(--dds-modal-container-padding);
 }
 
 .content {
   display: grid;
-  padding-right: var(--dds-spacing-x1);
+  padding-right: calc(var(--dds-spacing-x1) - var(--dds-modal-content-padding));
   gap: var(--dds-spacing-x1);
+}
+
+/* Spacing slik at fokusmarkering osv. ikke skjules ved overflow */
+.body {
+  padding: var(--dds-modal-content-padding);
+}
+
+.header {
+  padding: var(--dds-modal-content-padding);
 }
 
 .actions {
