--- packages/components/src/components/ScrollableContainer/ScrollableContainer.mdx
@@ -1,60 +1,0 @@
-import { Meta, ArgTypes, Canvas, Controls } from '@storybook/blocks';
-import { Scrollbar } from '.';
-import {
-  Source,
-  ComponentLinkRow,
-} from '@norges-domstoler/storybook-components';
-import * as ScrollableContainerStories from './ScrollableContainer.stories';
-
-<Meta of={ScrollableContainerStories} />
-
-# ScrollableContainer
-
-<ComponentLinkRow
-  docsHref="https://design.domstol.no/987b33f71/p/32db9d-scrollbar"
-  figmaHref="https://www.figma.com/file/ewqSDmkgqDQ5PyOsRp4V5b/DDS-Komponenter?node-id=7963%3A148079"
-  githubHref="https://github.com/domstolene/designsystem/tree/main/packages/components/src/components/ScrollableContainer"
-/>
-
-## Oversikt
-
-To subkomponenter er tilgjengelige:
-
-- **`<ScrollableContainer />`** - hovedkomponenten; en wrapper for innhold som skal scrolles. Håndterer scrolling og fokus ut av boksen.
-- **`<Scrollbar />`** - selve scrollbaren som brukes i `ScrollableContainer`. Håndterer scrolling når knyttet til elementet som skal scrolles. Kan brukes alene og håndtere plassering av scrollbaren og styling av innholdselementet selv.
-
-## Props
-
-### ScrollableContainer
-
-Hovedkomponenten for scrollbar innhold.
-
-<Canvas of={ScrollableContainerStories.Default} sourceState="shown" />
-<Controls of={ScrollableContainerStories.Default} />
-
-### Scrollbar
-
-Alenestående scrollbar til mer custom bruk.
-
-<ArgTypes of={Scrollbar} />
-
-## Bruk i koden
-
-### Vanlig bruk
-
-<Source
-  code={`import { ScrollableContainer } from '@norges-domstoler/dds-components';
-
-<ScrollableContainer>
-  <p>lang tekst</p>
-  <p>lang tekst</p>
-</ScrollableContainer>
-
-`}
-/>
-
-### Kun med `<Scrollbar />`
-
-Ved bruk av kun `<Scrollbar />` må man håndtere styling av containere, kobling til innholdet, plasseringen av `<Scrollbar/>` og fokus. Se koden i eksempelet.
-
-<Canvas of={ScrollableContainerStories.JustScrollbar} sourceState="shown" />
