--- packages/components/src/components/ScrollableContainer/index.ts
@@ -1,3 +1,0 @@
-export * from './Scrollbar';
-export * from './ScrollableContainer';
-export * from './scrollbarStyling';
