+++ packages/components/src/components/helpers/styling/index.ts
@@ -7,5 +7,6 @@ export * from './normalize';
 export * from './removeButtonStyling';
 export * from './removeListStyling';
 export * from './inheritLinkStyling';
+export * from './scrollbarStyling';
 export * from './selection';
 export * from './visibilityTransition';
