+++ packages/components/src/index.ts
@@ -36,7 +36,6 @@ export * from './components/OverflowMenu';
 export * from './components/Pagination';
 export * from './components/Popover';
 export * from './components/ProgressTracker';
-export * from './components/ScrollableContainer';
 export * from './components/Search';
 export * from './components/Select';
 export * from './components/SelectionControl/Checkbox';
